Revert "Feature/report function"
