+++ frontend/package.json
@@ -100,5 +100,5 @@
       "!src/apis/index.js"
     ]
   },
-  "proxy": "http://127.0.0.1:8000/"
+  "proxy": "http://localhost:8000"
 }
