--- frontend/src/components/posts/PostReportButton.jsx
@@ -1,75 +1,0 @@
-import React, { useState } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import styled from 'styled-components';
-import {
-  IconButton,
-  Card,
-  Grow,
-  List,
-  ListItem,
-  ListItemText,
-  Typography
-} from '@material-ui/core';
-import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
-
-const useStyles = makeStyles(() => ({
-  card: {
-    position: 'absolute',
-    right: '12px',
-    zIndex: 1
-  }
-}));
-
-const ReportButtonWrapper = styled.div`
-  justify-self: right;
-`;
-
-export default function PostReportButton() {
-  const classes = useStyles();
-  const [showButtons, setShowButtons] = useState(false);
-
-  const ItemText = ({ text }) => (
-    <Typography style={{ color: '#777', fontSize: 12 }}>{text}</Typography>
-  );
-
-  // TODO: 게시글 신고 기능 연결
-  const onClickReportPost = () => {};
-
-  // TODO: 사용자 신고 기능 연결
-  const onClickReportUser = () => {};
-
-  return (
-    <ReportButtonWrapper>
-      <IconButton
-        color="secondary"
-        id="report-button"
-        style={{ padding: '4px' }}
-        onClick={() => setShowButtons((prev) => !prev)}
-      >
-        <MoreHorizIcon className="more-button" />
-      </IconButton>
-      <Grow in={showButtons}>
-        <Card className={classes.card}>
-          <List style={{ padding: '0' }}>
-            <ListItem button>
-              <ListItemText
-                id="report-post-button"
-                primary={<ItemText text="게시글 신고" />}
-                onClick={onClickReportPost}
-              />
-            </ListItem>
-            <ListItem button>
-              <ListItemText
-                id="report-user-button"
-                primary={<ItemText text="사용자 신고" />}
-                onClick={onClickReportUser}
-              />
-            </ListItem>
-          </List>
-        </Card>
-      </Grow>
-    </ReportButtonWrapper>
-  );
-}
-
-PostReportButton.displayName = 'PostReportButton';
